Moved some files to the new utils directory.
