+++ TestRail/Types/Plan.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using Newtonsoft.Json.Linq;
+using TestRail.Utils;
 
 namespace TestRail.Types
 {
