+++ TestRail/Types/PlanEntry.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using Newtonsoft.Json.Linq;
+using TestRail.Utils;
 
 namespace TestRail.Types
 {
